Added some features.

+++ QoreLib/QoreLib/Models/SpectrumModel.cs
@@ -5,12 +5,12 @@ namespace QoreLib.Models;
 public class SpectrumModel
 {
     public int Id { get; set; }
-    public double[] DataX { get; set; }
-    public double[] DataY { get; set; }
-    public double[]? Omega01X { get; set; }
-    public double[]? Omega01Y { get; set; }
-    public double[]? Omega12X { get; set; }
-    public double[]? Omega12Y { get; set; }
+    public byte[] DataX { get; set; }
+    public byte[] DataY { get; set; }
+    public byte[]? Omega01X { get; set; }
+    public byte[]? Omega01Y { get; set; }
+    public byte[]? Omega12X { get; set; }
+    public byte[]? Omega12Y { get; set; }
     public string Type { get; set; }
     public string Name { get; set; }
     public bool IsFilled { get; set; }
